Shipping a phpstan.neon was pointless because Upgrade Status generates its own anyway.
Shipping rector.yml files is pointless since rector only supports PHP based config since 0.8: https://getrector.org/blog/2020/08/31/rector-is-moving-from-yaml-to-php-configs-what-changes-and-how-to-get-ready

This allowed to remove a hack from the analyzer. It also means we need to not do the test-only run for now, but that may not be that brittle anymore to be needed. We can put it back later with some work if needed.